pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "advision_env"
version = "1.0.0"
description = "AdVision OpenEnv Environment"
dependencies = [
    "openenv-core==0.2.3",
    "openai>=1.0.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "gymnasium>=0.29.0",
    "ultralytics>=8.0.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "stable-baselines3>=2.0.0",
    "opencv-python-headless>=4.8.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.100.0",
    "imageio>=2.37.3",
    "imageio-ffmpeg>=0.6.0",
    "requests>=2.28.0",
    "numpy>=1.24.0",
    "anthropic>=0.18.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["advision_env*"]

[project.scripts]
server = "server.app:main"
